chore: move all format utility functions to the separate formatters folder and update related imports
