+++ src/features/transactions/components/recent-transaction.tsx
@@ -2,8 +2,8 @@ import SpriteSvg from '@components/sprite-svg';
 import { Avatar, Box, colors, Typography } from '@mui/material';
 import { FC } from 'react';
 import { ITransaction } from '../types/transaction.types';
-import { formatDayMonthYear } from '@utils/format-day-month-year';
-import { formatCurrency } from '@utils/format-currency';
+import { formatDayMonthYear } from '@utils/formatters/format-day-month-year';
+import { formatCurrency } from '@utils/formatters/format-currency';
 
 interface RecentTransactionProps extends ITransaction {}
